Disable watching of the traefik config file
Watching /config/traefik.toml (which is on per default) would be desirable as it would automatically pick up changes in the corresponding configmap. However, having file watching enabled currently breaks the deployments that are autmatically created as a part of the travis CI/CD pipeline on the main Renku repository.
So far, all other deployments seem to be ok with the watch = "true" setting though...